repackage/clean up transform script code introduce AnalysisScript class to handle different scripting operations
