+++ api/src/org/labkey/api/assay/actions/AssayRunUploadForm.java
@@ -49,8 +49,8 @@ import org.labkey.api.exp.api.ExperimentService;
 import org.labkey.api.exp.property.Domain;
 import org.labkey.api.exp.property.DomainProperty;
 import org.labkey.api.pipeline.PipelineService;
-import org.labkey.api.qc.DefaultTransformResult;
-import org.labkey.api.qc.TransformResult;
+import org.labkey.api.assay.transform.DefaultTransformResult;
+import org.labkey.api.assay.transform.TransformResult;
 import org.labkey.api.qc.TsvDataExchangeHandler;
 import org.labkey.api.query.DefaultSchema;
 import org.labkey.api.query.PdLookupForeignKey;
