+++ api/src/org/labkey/api/assay/transform/DefaultTransformResult.java
@@ -1,0 +1,154 @@
+/*
+ * Copyright (c) 2009-2019 LabKey Corporation
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.labkey.api.assay.transform;
+
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
+import org.labkey.api.dataiterator.DataIteratorBuilder;
+import org.labkey.api.exp.api.ExpData;
+import org.labkey.api.exp.property.DomainProperty;
+import org.labkey.api.util.HtmlString;
+import org.labkey.vfs.FileLike;
+
+import java.io.File;
+import java.util.Collections;
+import java.util.List;
+import java.util.Map;
+
+/**
+ * User: klum
+ * Date: May 7, 2009
+ */
+public class DefaultTransformResult implements TransformResult
+{
+    private Map<ExpData, DataIteratorBuilder> _dataMap = Collections.emptyMap();
+    private Map<DomainProperty, String> _batchProperties = Collections.emptyMap();
+    private Map<DomainProperty, String> _runProperties = Collections.emptyMap();
+    private List<FileLike> _uploadedFiles;
+    private String _assayId;
+    private String _comments;
+    private HtmlString _warnings;
+    private List<File> _files;
+    private static final Logger LOG = LogManager.getLogger(DefaultTransformResult.class);
+
+    @Override
+    public List<File> getFiles()
+    {
+        return _files;
+    }
+
+    public void setFiles(List<File> files)
+    {
+        _files = files;
+    }
+
+    @Override
+    public HtmlString getWarnings()
+    {
+        return _warnings;
+    }
+
+    public void setWarnings(HtmlString warnings)
+    {
+        _warnings = warnings;
+    }
+
+    public DefaultTransformResult(){}
+
+    public DefaultTransformResult(Map<ExpData, DataIteratorBuilder> dataMap)
+    {
+        _dataMap = dataMap;
+        _assayId = null;
+    }
+
+    public DefaultTransformResult(TransformResult mergeResult)
+    {
+        _dataMap = mergeResult.getTransformedData();
+        _batchProperties = mergeResult.getBatchProperties();
+        _runProperties = mergeResult.getRunProperties();
+        _uploadedFiles = mergeResult.getUploadedFiles();
+        _warnings = mergeResult.getWarnings();
+        _files = mergeResult.getFiles();
+        _assayId = null;
+    }
+
+    @Override
+    public Map<ExpData, DataIteratorBuilder> getTransformedData()
+    {
+        return _dataMap;
+    }
+
+    @Override
+    public Map<DomainProperty, String> getBatchProperties()
+    {
+        return _batchProperties;
+    }
+
+    public void setBatchProperties(Map<DomainProperty, String> batchProperties)
+    {
+        _batchProperties = batchProperties;
+    }
+
+    @Override
+    public Map<DomainProperty, String> getRunProperties()
+    {
+        return _runProperties;
+    }
+
+    public void setRunProperties(Map<DomainProperty, String> runProperties)
+    {
+        _runProperties = runProperties;
+    }
+
+    @Override
+    public List<FileLike> getUploadedFiles()
+    {
+        return _uploadedFiles;
+    }
+
+    public void setUploadedFiles(List<FileLike> uploadedFiles)
+    {
+        _uploadedFiles = uploadedFiles;
+    }
+
+    public static TransformResult createEmptyResult()
+    {
+        return new DefaultTransformResult();
+    }
+
+    @Override
+    public String getAssayId()
+    {
+        return _assayId;
+    }
+
+    public void setAssayId(String assayId)
+    {
+        _assayId = assayId;
+    }
+
+    @Override
+    public String getComments()
+    {
+        return _comments;
+    }
+
+    public void setComments(String comments)
+    {
+        _comments = comments;
+    }
+}
