+++ api/src/org/labkey/api/assay/transform/TransformResult.java
@@ -1,0 +1,64 @@
+/*
+ * Copyright (c) 2009-2019 LabKey Corporation
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.labkey.api.assay.transform;
+
+import org.labkey.api.dataiterator.DataIteratorBuilder;
+import org.labkey.api.exp.api.ExpData;
+import org.labkey.api.exp.property.DomainProperty;
+import org.labkey.api.util.HtmlString;
+import org.labkey.vfs.FileLike;
+
+import java.io.File;
+import java.util.List;
+import java.util.Map;
+
+/**
+ * User: klum
+ * Date: May 7, 2009
+ */
+public interface TransformResult
+{
+    Map<ExpData, DataIteratorBuilder> getTransformedData();
+
+    Map<DomainProperty, String> getRunProperties();
+
+    Map<DomainProperty, String> getBatchProperties();
+
+    List<FileLike> getUploadedFiles();
+
+    String getAssayId();
+
+    String getComments();
+
+    /**
+     * Get warnings generated by the transform script. Warnings saved if run properties file severity level
+     * is set to WARN and transform script writes to errors.html in its working directory.  Transform script must
+     * also set maximum severity in transform run properties to WARN.
+     *
+     * @return      the string containing the error message
+     */
+    HtmlString getWarnings();
+
+    /**
+     * Get transform script output files.  These files are collected from the transform scripts operating directory and
+     * include stdout and stderr piped to file. The output data files, indicated by runDataUploadedFile in the run
+     * properties file, will also be included.
+     *
+     * @return      List of files gathered
+     */
+    List<File> getFiles();
+}
