+++ api/src/org/labkey/api/qc/TsvDataExchangeHandler.java
@@ -31,6 +31,9 @@ import org.labkey.api.assay.DefaultAssayRunCreator;
 import org.labkey.api.assay.TsvDataHandler;
 import org.labkey.api.assay.actions.AssayRunUploadForm;
 import org.labkey.api.assay.actions.ProtocolIdForm;
+import org.labkey.api.assay.transform.DataExchangeHandler;
+import org.labkey.api.assay.transform.DefaultTransformResult;
+import org.labkey.api.assay.transform.TransformResult;
 import org.labkey.api.collections.CaseInsensitiveHashMap;
 import org.labkey.api.data.Container;
 import org.labkey.api.data.TSVWriter;
@@ -632,7 +635,6 @@ public class TsvDataExchangeHandler implements DataExchangeHandler
                 throw new ValidationException("Transform script has thrown errors.");
             }
         }
-
     }
 
     public void processValidationOutput(RunInfo info, @Nullable Logger log) throws ValidationException
