+++ assay/src/org/labkey/assay/AssayDomainServiceImpl.java
@@ -32,6 +32,8 @@ import org.labkey.api.assay.plate.Plate;
 import org.labkey.api.assay.plate.PlateBasedAssayProvider;
 import org.labkey.api.assay.plate.PlateService;
 import org.labkey.api.assay.security.DesignAssayPermission;
+import org.labkey.api.assay.transform.AnalysisScript;
+import org.labkey.api.assay.transform.DataTransformService;
 import org.labkey.api.collections.CaseInsensitiveHashSet;
 import org.labkey.api.data.Container;
 import org.labkey.api.data.ContainerManager;
@@ -247,12 +249,12 @@ public class AssayDomainServiceImpl extends BaseRemoteService implements AssayDo
             result.setAvailableDetectionMethods(dmProvider.getAvailableDetectionMethods());
         }
 
-        List<File> typeScripts = provider.getValidationAndAnalysisScripts(protocol, AssayProvider.Scope.ASSAY_TYPE);
+        List<AnalysisScript> typeScripts = provider.getValidationAndAnalysisScripts(protocol, AssayProvider.Scope.ASSAY_TYPE);
         if (!typeScripts.isEmpty())
         {
             List<String> scriptNames = new ArrayList<>();
-            for (File script : typeScripts)
-                scriptNames.add(script.getAbsolutePath());
+            for (AnalysisScript script : typeScripts)
+                scriptNames.add(script.getScript().toNioPathForRead().toString());
 
             result.setModuleTransformScripts(scriptNames);
         }
@@ -264,12 +266,13 @@ public class AssayDomainServiceImpl extends BaseRemoteService implements AssayDo
         result.setPlateMetadata(provider.isPlateMetadataEnabled(protocol));
 
         // data transform scripts
-        List<File> transformScripts = provider.getValidationAndAnalysisScripts(protocol, AssayProvider.Scope.ASSAY_DEF);
+        List<AnalysisScript> transformScripts = provider.getValidationAndAnalysisScripts(protocol, AssayProvider.Scope.ASSAY_DEF);
 
         List<String> transformScriptStrings = new ArrayList<>();
-        for (File transformScript : transformScripts)
+        for (AnalysisScript transformScript : transformScripts)
         {
-            transformScriptStrings.add(transformScript.getAbsolutePath());
+            // TODO, add allowable operations once we have UI that controls those options
+            transformScriptStrings.add(transformScript.getScript().toNioPathForRead().toString());
         }
         result.setProtocolTransformScripts(transformScriptStrings);
 
@@ -490,7 +493,7 @@ public class AssayDomainServiceImpl extends BaseRemoteService implements AssayDo
                 }
 
                 // data transform scripts
-                List<File> transformScripts = new ArrayList<>();
+                List<AnalysisScript> transformScripts = new ArrayList<>();
                 List<String> submittedScripts = assay.getProtocolTransformScripts();
                 if (!submittedScripts.isEmpty() && !canUpdateTransformationScript())
                     throw new ValidationException("You must be a platform developer or site admin to configure assay transformation scripts.");
@@ -498,7 +501,8 @@ public class AssayDomainServiceImpl extends BaseRemoteService implements AssayDo
                 {
                     if (!StringUtils.isBlank(script))
                     {
-                        transformScripts.add(new File(script));
+                        // TODO : handle analysis script operations once UI is created
+                        transformScripts.add(new AnalysisScript(new File(script), Set.of(DataTransformService.TransformOperation.INSERT)));
                     }
                 }
